% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsb-asinfo.r
\name{gsb_asinfo}
\alias{gsb_asinfo}
\title{Retrive AS info from Google SafeBrowsing}
\usage{
gsb_asinfo(as_size = c("largest", "all"), time_range = 90,
  type_detected = c("both", "attack", "compromised"), region = "all",
  .progress = FALSE)
}
\arguments{
\item{as_size}{return results for \code{all} ASNs or just the \code{largest}?}

\item{time_range}{number of days of history for the time series}

\item{type_detected}{one of \code{both}, \code{attack} or \code{compromised}.}

\item{region}{ISO2 region code or \code{all}}

\item{.progress}{display a progress bar?}
}
\value{
\code{data.frame} (\code{tbl_df}) with ASN info
}
\description{
Retrive AS info from Google SafeBrowsing
}
\note{
Depending on the parameters used (especially \code{time_range}),
      this opereation could take a while. You should probably turn on
      progress bars for any query > 7 days.
}
\examples{
\dontrun{
gsb_asinfo("largest", 7)
}
}
\references{
See \href{https://www.google.com/transparencyreport/safebrowsing/malware/?hl=en}{Google's page}
      for more information on how they scan & report ASN info.
}

