% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site-status.r
\name{gsb_site_status}
\alias{gsb_site_status}
\title{Retrive URL "site status"}
\usage{
gsb_site_status(site_urls)
}
\arguments{
\item{site_url}{non-relative URL to check}
}
\value{
\code{data.frame} (\code{tbl_df}) with site status
}
\description{
Pass in a vector of URLs and the Google Safe Browsing Site Status
for the URLs will be returned in a \code{data.frame} (\code{tbl_df})
}
\examples{
\dontrun{
gsb_site_status(c("http://fgcdiesel.cl/encrypted.exe",
                  "http://rud.is/",
                  "http://dds.ec/"))
}
}

